requirements.txt
backcall==0.2.0
decorator==5.1.0
ipython==7.30.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
Pygments==2.10.0
PyYAML==6.0
traitlets==5.1.1
wcwidth==0.2.5
xmltodict==0.12.0
